+++ packages/vexide-devices/src/smart/serial.rs
@@ -83,6 +83,24 @@ impl SerialPort {
         Self { port, device }
     }
 
+    /// Configures the baud rate of the serial port.
+    ///
+    /// Baud rate determines the speed of communication over the data channel. Under normal conditions, user code is limited
+    /// to a maximum baudrate of 921600.
+    ///
+    /// # Errors
+    ///
+    /// - A [`SerialError::Port`] error is returned if a generic serial device is not currently connected to the Smart Port.
+    pub fn set_baud_rate(&mut self, baud_rate: u32) -> Result<(), SerialError> {
+        self.validate_port()?;
+
+        unsafe {
+            vexDeviceGenericSerialBaudrate(self.device, baud_rate as i32);
+        }
+
+        Ok(())
+    }
+
     /// Clears the internal input and output FIFO buffers.
     ///
     /// This can be useful to reset state and remove old, potentially unneeded data
@@ -148,7 +166,11 @@ impl SerialPort {
     ///     }
     /// }
     /// ```
-    pub fn read_byte(&self) -> Result<Option<u8>, SerialError> {
+    ///
+    /// # Errors
+    ///
+    /// - A [`SerialError::Port`] error is returned if a generic serial device is not currently connected to the Smart Port.
+    pub fn read_byte(&mut self) -> Result<Option<u8>, SerialError> {
         self.validate_port()?;
 
         let byte = unsafe { vexDeviceGenericSerialReadChar(self.device) };